% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagged_svm.R
\name{add_baggedSVM_engine}
\alias{add_baggedSVM_engine}
\title{Wrapper to add a `baggedSVM` engine to the parsnip svm_rbf model
specification}
\usage{
add_baggedSVM_engine()
}
\description{
Wrapper to add a `baggedSVM` engine to the parsnip svm_rbf model
specification
}
