% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree_cubist.R
\name{add_cubist_engine}
\alias{add_cubist_engine}
\title{Wrapper to add the `Cubist` engine to the parsnip `boost_tree` models
specification}
\usage{
add_cubist_engine()
}
\description{
Wrapper to add the `Cubist` engine to the parsnip `boost_tree` models
specification
}
