% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastNaiveBayes.R
\name{add_fastNaiveBayes_engine}
\alias{add_fastNaiveBayes_engine}
\title{Wrapper to add the `fastNaiveBayes` engine to the parsnip `naive_Bayes` model
specification in the discrim package}
\usage{
add_fastNaiveBayes_engine()
}
\description{
Wrapper to add the `fastNaiveBayes` engine to the parsnip `naive_Bayes` model
specification in the discrim package
}
