% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnn.R
\name{add_fnn_engine}
\alias{add_fnn_engine}
\title{Wrapper to add the `FNN` engine to the parsnip `nearest_neighbor` model
specification}
\usage{
add_fnn_engine()
}
\description{
Wrapper to add the `FNN` engine to the parsnip `nearest_neighbor` model
specification
}
