% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_h2o.R
\name{add_h2o_engine}
\alias{add_h2o_engine}
\title{Wrapper to add the `h2o` engine to the parsnip `mlp` model specification}
\usage{
add_h2o_engine()
}
\description{
Wrapper to add the `h2o` engine to the parsnip `mlp` model specification
}
