% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_liquidsvm.R
\name{add_liquidSVM_engine}
\alias{add_liquidSVM_engine}
\title{Wrapper to add the `liquidSVM` engine to the parsnip svm_rbf model
specification}
\usage{
add_liquidSVM_engine()
}
\description{
Wrapper to add the `liquidSVM` engine to the parsnip svm_rbf model
specification
}
