% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralnet.R
\name{add_neuralnet_engine}
\alias{add_neuralnet_engine}
\title{Wrapper to add the `neuralnet` engine to the parsnip `mlp` model
specification}
\usage{
add_neuralnet_engine()
}
\description{
Wrapper to add the `neuralnet` engine to the parsnip `mlp` model
specification
}
