% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rknn.R
\name{add_rknn_engine}
\alias{add_rknn_engine}
\title{Wrapper to add the `rknn` engine to the parsnip `nearest_neighbor` model
specification}
\usage{
add_rknn_engine()
}
\description{
Wrapper to add the `rknn` engine to the parsnip `nearest_neighbor` model
specification
}
