% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagged_svm.R
\name{baggedSVM_predict}
\alias{baggedSVM_predict}
\title{Prediction wrapper for baggedSVM}
\usage{
baggedSVM_predict(object, newdata, type, ...)
}
\arguments{
\item{object}{A fitted model}

\item{newdata}{A data.frame}

\item{type}{A character specifying the prediction type, either "class",
"numeric"}

\item{...}{Currently unused}
}
\value{
Prediction results
}
\description{
Prediction wrapper for baggedSVM
}
