% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagged_svm.R
\name{baggedSVM_train}
\alias{baggedSVM_train}
\title{Training wrapper for baggedSVM}
\usage{
baggedSVM_train(
  formula,
  data,
  mode,
  cost = 1,
  rbf_sigma = 0.05,
  margin = 0.1,
  ...
)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A data.frame of training data.}

\item{mode}{A character specifying "classification" or "regression".}

\item{cost}{The dials::cost parmaeters.}

\item{rbf_sigma}{The dials::rbf_sigma parameter.}

\item{margin}{The dials::margin parameter.}

\item{...}{Currently unused.}

\item{nfolds}{An engine argument used to set the number of folds for bagging.}
}
\value{
A fitted parsnip model specification.
}
\description{
Training wrapper for baggedSVM
}
