% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging.R
\name{bagging}
\alias{bagging}
\title{General interface for bagging models}
\usage{
bagging(
  mode = "classification",
  base_model = NULL,
  mtry = NULL,
  trees = 10,
  sample_prop = 0.67
)
}
\arguments{
\item{mode}{A single character string for the type of model.}

\item{base_model}{The parsnip model specification to use as the base model in
the bagging ensemble.}

\item{mtry}{The number of predictors that will be randomly sampled for each
model in the bagging ensemble.}

\item{trees}{The number of bagging models in the ensemble.}

\item{sample_prop}{The number or proportion of samples to use in each bagging
model.}

\item{model}{A parsnip `model_spec` object for use in the bagging ensemble.}
}
\value{
A model_spec
}
\description{
General interface for bagging models
}
