% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging_data.R
\name{bagging_predict}
\alias{bagging_predict}
\title{Prediction method for bagging}
\usage{
bagging_predict(object, newdata, type, ...)
}
\arguments{
\item{object}{A fitted model}

\item{newdata}{A data.frame}

\item{type}{A character specifying the prediction type, either "class",
"numeric"}

\item{...}{Currently unused}
}
\value{
Prediction results
}
\description{
Prediction method for bagging
}
