% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagging_data.R
\name{bagging_train}
\alias{bagging_train}
\title{Training method for bagging}
\usage{
bagging_train(
  formula,
  data,
  mode,
  mtry = NULL,
  trees = 10,
  sample_prop = 0.67,
  model = NULL,
  .opts = furrr::future_options(),
  ...
)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{A data.frame of training data.}

\item{mode}{A character specifying "classification" or "regression".}

\item{mtry}{The dials::mtry parameters.}

\item{trees}{The dials::trees parameter.}

\item{sample_prop}{The dials::sample_prop parameter.}

\item{model}{The base model specification to use in the bagging ensemble.}

\item{...}{Currently unused.}

\item{.opt}{Additional future_options to use for multiprocessing.}
}
\value{
A fitted parsnip model specification.
}
\description{
Training method for bagging
}
