% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubist_parsnip.R
\name{cubist_forest}
\alias{cubist_forest}
\title{Cubist regression trees model specification}
\usage{
cubist_forest(mode = "unknown", trees = 1, neighbors = 0)
}
\arguments{
\item{mode}{character, one of c("classification", "regression")}

\item{trees}{integer, number of boosting iterations}

\item{neighbors}{integer, number of neighbors}
}
\value{
model_spec object
}
\description{
Cubist regression trees model specification
}
