% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{General interface for a stacked ensemble of models}
\usage{
ensemble(mode = "unknown")
}
\arguments{
\item{mode}{A single character string for the type of model.}
}
\value{
A model_spec
}
\description{
General interface for a stacked ensemble of models
}
