% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_predict}
\alias{ensemble_predict}
\title{Prediction method for ensembles}
\usage{
ensemble_predict(object, newdata, type = NULL, ...)
}
\arguments{
\item{object}{A model_fit object}

\item{newdata}{A data.frame}

\item{type}{Either "regression" or "classification"}

\item{...}{Other arguments currently unused}
}
\value{
A tibble
}
\description{
Prediction method for ensembles
}
