% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_train}
\alias{ensemble_train}
\title{Training method for a stacked ensemble}
\usage{
ensemble_train(
  formula,
  data,
  mode,
  base_workflows,
  meta_model,
  cv = purrr::partial(rsample::vfold_cv, v = 5),
  ...
)
}
\arguments{
\item{formula}{A formula}

\item{data}{A data.frame to train on}

\item{mode}{Either "regression" or "classification"}

\item{base_workflows}{A named list of workflows to use as base models.}

\item{meta_model}{A parsnip model specification to use as the meta model.}

\item{cv}{A rsample resamples object. Default uses 5-fold cross validation.}

\item{...}{Other arguments currently unused.}
}
\value{
A model_fit object
}
\description{
Training method for a stacked ensemble
}
