% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighor_fnn.R
\name{fnn_pred}
\alias{fnn_pred}
\title{Nearest neighbors prediction using FNN}
\usage{
fnn_pred(object, newdata, prob = FALSE, ...)
}
\arguments{
\item{object}{parsnip model spec.}

\item{newdata}{data.frame or matrix of training data.}

\item{prob}{logical return predicted probability of the winning class,
default = FALSE.}

\item{...}{additional arguments to pass to FNN, currently unused.}
}
\value{
data.frame containing the predicted results.
}
\description{
`fnn_pred` is a wrapper for `FNN` fast nearest neighbor models
}
