% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_h2o.R
\name{h2o_train}
\alias{h2o_train}
\title{Parsnip model specification}
\usage{
h2o_train(
  formula,
  data,
  l2 = 0,
  hidden_dropout_ratios = 0,
  hidden = 200,
  epochs = 10,
  activation = "RectifierWithDropout",
  ...
)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame of training data}

\item{l2}{numeric, l2 regulation parameter, default = 0}

\item{hidden_dropout_ratios}{dropout ratio for single hidden layer, default
= 0}

\item{hidden}{integer, number of neurons in the hidden layer, default = 200}

\item{epochs}{integer, number of epochs, default = 10}

\item{activation}{character, activation function, default is 
"RectifierWithDropout"}

\item{...}{other arguments currently unused}
}
\value{
evaluated h2o model call
}
\description{
Parsnip model specification
}
