% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_predict}
\alias{kriging_predict}
\title{Predict function for regression kriging}
\usage{
kriging_predict(object, new_data, type, nmax, ...)
}
\arguments{
\item{object}{A kriging_rk class object.}

\item{new_data}{An sf object or a SpatialGridDataFrame.}

\item{type}{A character with either "numeric" or "conf_int".}

\item{nmax}{Integer, number of closest points to use in prediction.}

\item{...}{Other arguments passed to gstat::krige function.}
}
\value{

}
\description{
Predict function for regression kriging
}
