% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_rk}
\alias{kriging_rk}
\title{Regression kriging model specification}
\usage{
kriging_rk(mode = "regression", neighbors = NULL, nscore = FALSE)
}
\arguments{
\item{mode}{A character specifying the mode of the model. Only 'regression'
is currently implemented.}

\item{neighbors}{An integer specifying the maximum number of points to use in
neighborhood.}

\item{nscore}{A logical to optionally transform and backtransform data using
a normal score transform. Default is FALSE.}
}
\description{
Regression kriging model specification
}
