% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_train}
\alias{kriging_train}
\title{Regression Kriging}
\usage{
kriging_train(formula, data, ...)
}
\arguments{
\item{...}{arguments to pass onto the automap::autofitVariogram function}

\item{object}{formula}

\item{input_data}{sf object containing sfc points features}
}
\value{
kriging_rk class object
}
\description{
Regression Kriging
}
