% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{multi_predict._kriging_rk}
\alias{multi_predict._kriging_rk}
\title{`multi_predict`` method for krige_rk}
\usage{
\method{multi_predict}{`_kriging_rk`}(object, new_data, type = "numeric", neighbors = 16, ...)
}
\arguments{
\item{object}{A model_spec.}

\item{new_data}{A data.frame.}

\item{type}{A character specifying either "numeric" or "conf_int".}

\item{neighbors}{An integer specifying number of closest points to use in
prediction.}

\item{...}{Currently unused.}
}
\value{
A tibble with the prediction results.
}
\description{
`multi_predict`` method for krige_rk
}
