% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{print.kriging_rk}
\alias{print.kriging_rk}
\title{Print generic function method for regression kriging model}
\usage{
\method{print}{kriging_rk}(x, ...)
}
\arguments{
\item{x}{A model object}

\item{...}{Not currently used}
}
\description{
Print generic function method for regression kriging model
}
