% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_liquid.R
\name{read_liquid}
\alias{read_liquid}
\title{Reads a liquidSVM model and inserts it back into a workflow or parsnip model
specification that is based on the liquidSVM engine.}
\usage{
read_liquid(object, filename, ...)
}
\arguments{
\item{filename}{The file path to the saved model.}

\item{...}{Currently unused.}

\item{wrkflow}{A workflow object or a parsnip `model_fit` that is based on
the liquidSVM engine.}
}
\value{
A workflow object with the restored liquidSVM fit.
}
\description{
Reads a liquidSVM model and inserts it back into a workflow or parsnip model
specification that is based on the liquidSVM engine.
}
