% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rknn.R
\name{rknn_pred}
\alias{rknn_pred}
\title{Nearest neighbors prediction using rknn}
\usage{
rknn_pred(object, newdata, ...)
}
\arguments{
\item{object}{parsnip model spec.}

\item{newdata}{data.frame or matrix of training data.}

\item{...}{additional arguments to pass to FNN, currently unused.}
}
\value{
data.frame containing the predicted results.
}
\description{
`rknn_pred` is a wrapper for `rknn` fast nearest neighbor models
}
