% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rknn.R
\name{rknn_train}
\alias{rknn_train}
\title{Nearest neighbors using rknn}
\usage{
rknn_train(x, y = NULL, k = 1, r = 500, mtry = trunc(sqrt(ncol(x))), ...)
}
\arguments{
\item{x}{a data frame or matrix of predictors.}

\item{y}{a vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{k}{a vector (integer) of the number of neighbours to consider.}

\item{r}{an integer specifying the number of KNN models containing random
feature subsets. Default = 500.}

\item{mtry}{an integer with the number of features to randomly select per KNN
model. Default is `trunc(sqrt(ncol(x)))`.}

\item{...}{additional arguments to pass to FNN, currently unused.}
}
\value{
list containing the FNN call
}
\description{
`rknn_train` is a wrapper for `rknn` fast nearest neighbor models
}
