% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_liquid.R
\name{svm_liquid}
\alias{svm_liquid}
\title{liquidSVM model}
\usage{
svm_liquid(mode = "unknown", cost = NULL, rbf_sigma = NULL,
  margin = NULL)
}
\arguments{
\item{mode}{character, one of c("classification", "regression")}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{rbf_sigma}{A positive number for radial basis function.}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\value{
model_spec object
}
\description{
liquidSVM model
}
