% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_liquidsvm.R
\name{train_liquid}
\alias{train_liquid}
\title{Wrapper to train the `liquidSVM` engine}
\usage{
train_liquid(x, y, lambdas, gammas, ...)
}
\arguments{
\item{x}{A matrix of training data/}

\item{y}{A vector of response data.}

\item{lambdas}{A numeric with the amount of regularization.}

\item{gammas}{A numeric with the bandwidth of the gaussian kernel.}

\item{...}{Other engine arguments.}
}
\value{

}
\description{
Wrapper to train the `liquidSVM` engine
}
