% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_liquid.R
\name{write_liquid}
\alias{write_liquid}
\title{Saves a liquidSVM model that is contained within a parsnip model
specification or a workflows object}
\usage{
write_liquid(object, filename, ...)
}
\arguments{
\item{filename}{The file path to save the model.}

\item{...}{Currently unused.}

\item{wrkflow}{A parsnip `model_fit` or workflow object based on the
liquidSVM engine.}
}
\value{
None
}
\description{
Saves a liquidSVM model that is contained within a parsnip model
specification or a workflows object
}
