% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_predict}
\alias{kriging_predict}
\title{Predict function for regression kriging}
\usage{
kriging_predict(object, new_data, type, ...)
}
\arguments{
\item{object}{kriging_rk class object.}

\item{new_data}{sf object, or SpatialGridDataFrame.}

\item{type}{"numeric" or "conf_int".}

\item{...}{other arguments passed to gstat::krige function}
}
\value{

}
\description{
Predict function for regression kriging
}
