% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_rk}
\alias{kriging_rk}
\title{Regression kriging model specification}
\usage{
kriging_rk(mode = "regression", neighbors = NULL, nscore = FALSE)
}
\arguments{
\item{mode}{character, only 'regression' is currently implemented}

\item{neighbors}{integer, maximum number of points to use in neighborhood}

\item{nscore, }{logical optionally transform and backtransform data using a
normal score transform}
}
\description{
Regression kriging model specification
}
