% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_train}
\alias{kriging_train}
\title{Regression Kriging}
\usage{
kriging_train(formula, data, nscore = FALSE, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{sf object containing sfc points features}

\item{nscore}{logical, whether to transform the training data using a normal
score transform.}

\item{...}{arguments to pass onto the automap::autofitVariogram function}
}
\value{
kriging_rk class object
}
\description{
Regression Kriging
}
