% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{kriging_train}
\alias{kriging_train}
\title{Regression Kriging}
\usage{
kriging_train(formula, data, nmax, nscore = FALSE, ...)
}
\arguments{
\item{formula}{A formula object.}

\item{data}{An data.frame object that includes "x" and "y" columns.}

\item{nmax}{Integer, number of neighbors to use in prediction.}

\item{nscore}{A logical, whether to transform the training data using a normal
score transform.}

\item{...}{Other arguments to pass onto the automap::autofitVariogram
function.}
}
\value{
kriging_rk class object
}
\description{
Regression Kriging
}
