% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{multi_predict._kriging_rk}
\alias{multi_predict._kriging_rk}
\title{Multi_predict method for regression kriging model specification}
\usage{
multi_predict._kriging_rk(
  object,
  new_data,
  type = "numeric",
  neighbors = 16,
  ...
)
}
\arguments{
\item{object}{}

\item{new_data}{}

\item{type}{}

\item{...}{}

\item{nmax}{}
}
\value{

}
\description{
Multi_predict method for regression kriging model specification
}
