% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R, R/nscore.R
\name{nscore}
\alias{nscore}
\title{Parameter for logical parameter for normal score transform}
\usage{
nscore(x)

nscore(x)
}
\arguments{
\item{x}{numeric, vector}
}
\value{


list
nscore contains normal score transformed vector
trn.table contains an ordered table of the original values for
use with backtransformation
}
\description{
Takes a vector of values x and calculates their normal scores. Returns
}
