% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{update.kriging_rk}
\alias{update.kriging_rk}
\title{Method to update a regression kriging model specification}
\usage{
\method{update}{kriging_rk}(object, parameters = NULL, neighbors = NULL, fresh = FALSE, ...)
}
\arguments{
\item{object}{A model_spec.}

\item{parameters}{The parameters for the model}

\item{neighbors}{An integer specifying the maximum number of points to use in
neighborhood.}

\item{fresh}{A logical.}

\item{...}{Currently unused.}
}
\description{
Method to update a regression kriging model specification
}
