% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging_gstat.R
\name{update.kriging_rk}
\alias{update.kriging_rk}
\title{Method to update a regression kriging model specification}
\usage{
\method{update}{kriging_rk}(
  object,
  parameters = NULL,
  neighbors = NULL,
  nscore = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{model_spec}

\item{parameters}{parameters}

\item{neighbors}{integer}

\item{nscore}{logical}

\item{fresh}{logical}

\item{...}{currently unused}
}
\description{
Method to update a regression kriging model specification
}
